# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/search": {
      "get": {
        "operationId": "SearchController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
    "/search/karaoke": {
      "get": {
        "operationId": "SearchController_searchByKaraokeNo",
        "parameters": [
          {
            "name": "no",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "provider",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
    "/search/artist": {
      "get": {
        "operationId": "SearchController_searchByArtist",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
    "/search/title": {
      "get": {
        "operationId": "SearchController_searchByTitle",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Search"
        ]
      }
    },
    "/artists": {
      "get": {
        "operationId": "ArtistsController_findAll",
        "parameters": [
          {
            "name": "includeYoutube",
            "required": false,
            "in": "query",
            "description": "YouTube 정보 포함 여부 (true/false)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "아티스트 목록"
          }
        },
        "summary": "아티스트 목록 조회",
        "tags": [
          "Artists"
        ]
      }
    },
    "/artists/{identifier}": {
      "get": {
        "operationId": "ArtistsController_findByIdOrAlias",
        "parameters": [
          {
            "name": "identifier",
            "required": true,
            "in": "path",
            "description": "아티스트 ID 또는 alias",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "아티스트 상세 정보"
          },
          "404": {
            "description": "아티스트를 찾을 수 없음"
          }
        },
        "summary": "아티스트 상세 조회 (ID 또는 alias)",
        "tags": [
          "Artists"
        ]
      }
    },
    "/songs": {
      "get": {
        "operationId": "SongsController_findAll",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "검색어 (제목으로 검색)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "artistId",
            "required": false,
            "in": "query",
            "description": "아티스트 ID로 필터링",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "곡 목록"
          }
        },
        "summary": "곡 목록 조회 (검색 및 필터링)",
        "tags": [
          "Songs"
        ]
      }
    },
    "/songs/{id}": {
      "get": {
        "operationId": "SongsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "곡 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "곡 상세 정보"
          },
          "404": {
            "description": "곡을 찾을 수 없음"
          }
        },
        "summary": "곡 상세 조회",
        "tags": [
          "Songs"
        ]
      }
    },
    "/youtube": {
      "get": {
        "operationId": "YoutubeController_getOembedData",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "YouTube 동영상 URL",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OEmbed 데이터"
          },
          "400": {
            "description": "URL 파라미터 필요"
          }
        },
        "summary": "YouTube 동영상 OEmbed 데이터 조회",
        "tags": [
          "YouTube"
        ]
      }
    },
    "/youtube/search-channels": {
      "get": {
        "operationId": "YoutubeController_searchChannels",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "description": "아티스트 이름",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 결과"
          },
          "400": {
            "description": "아티스트 이름 필요"
          }
        },
        "summary": "YouTube 채널 검색",
        "tags": [
          "YouTube"
        ]
      }
    },
    "/youtube/update-artist-channel/{alias}": {
      "post": {
        "operationId": "YoutubeController_updateArtistChannel",
        "parameters": [
          {
            "name": "alias",
            "required": true,
            "in": "path",
            "description": "아티스트 alias",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "query",
            "description": "YouTube 채널 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "업데이트 성공"
          },
          "400": {
            "description": "channelId 필요"
          }
        },
        "summary": "아티스트 YouTube 채널 정보 업데이트",
        "tags": [
          "YouTube"
        ]
      }
    },
    "/scrape/blog/{id}": {
      "get": {
        "operationId": "BlogScrapeController_scrapeBlog",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BlogScrape"
        ]
      }
    },
    "/admin/artists": {
      "get": {
        "operationId": "AdminController_getArtists",
        "parameters": [],
        "responses": {
          "200": {
            "description": "아티스트 목록"
          }
        },
        "summary": "아티스트 목록 조회 (관리자용, 곡 수 포함)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/artists/{id}/songs": {
      "get": {
        "operationId": "AdminController_getArtistSongs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "아티스트 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "곡 목록"
          }
        },
        "summary": "특정 아티스트의 곡 목록 조회 (관리자용)",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "노래방 검색 API",
    "description": "노래방 검색 서비스 API 문서",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {}
  }
}